:recycle: refactor : company package separation, move file, add pageable size = 10, 30, 50
