+++ com.rush.logistic.client.company-product/src/main/java/com/rush/logistic/client/company_product/domain/company/service/CompanyService.java
@@ -1,21 +1,23 @@
-package com.rush.logistic.client.company_product.domain.service;
+package com.rush.logistic.client.company_product.domain.company.service;
 
-import com.rush.logistic.client.company_product.domain.dto.CompanyDto;
-import com.rush.logistic.client.company_product.domain.dto.request.CompanyCreateRequest;
-import com.rush.logistic.client.company_product.domain.dto.request.CompanyUpdateRequest;
-import com.rush.logistic.client.company_product.domain.dto.response.CompanySearchResponse;
-import com.rush.logistic.client.company_product.domain.entity.Company;
-import com.rush.logistic.client.company_product.domain.repository.CompanyRepository;
+import com.rush.logistic.client.company_product.domain.company.dto.CompanyDto;
+import com.rush.logistic.client.company_product.domain.company.dto.request.CompanyCreateRequest;
+import com.rush.logistic.client.company_product.domain.company.dto.request.CompanyUpdateRequest;
+import com.rush.logistic.client.company_product.domain.company.dto.response.CompanySearchResponse;
+import com.rush.logistic.client.company_product.domain.company.entity.Company;
+import com.rush.logistic.client.company_product.domain.company.repository.CompanyRepository;
 import com.rush.logistic.client.company_product.global.exception.ApplicationException;
 import com.rush.logistic.client.company_product.global.exception.ErrorCode;
 import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
+import java.util.Arrays;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -47,6 +49,15 @@ public class CompanyService {
     //업체 전체 조회
     @Transactional
     public Page<CompanyDto> getAllCompany(Pageable pageable) {
+        // 페이지 사이즈 제한
+        int[] allowedPageSizes = {10, 30, 50};
+        int pageSize = pageable.getPageSize();
+
+        // 허용되지 않은 페이지 사이즈일 경우 기본값 10으로 설정
+        if (!Arrays.stream(allowedPageSizes).anyMatch(size -> size == pageSize)) {
+            pageable = PageRequest.of(pageable.getPageNumber(), 10, pageable.getSort());
+        }
+
         Page<Company> companies = companyRepository.findAll(pageable);
         return companies.map(CompanyDto::from);
     }
