+++ com.rush.logistic.client.company-product/src/main/java/com/rush/logistic/client/company_product/global/type/CompanyType.java
@@ -1,4 +1,4 @@
-package com.rush.logistic.client.company_product.domain.type;
+package com.rush.logistic.client.company_product.global.type;
 
 public enum CompanyType {
     MANUFACTURER,  // 생산
